+++ src/Temporalio/Worker/WorkflowWorkerOptions.cs
@@ -23,6 +23,7 @@ namespace Temporalio.Worker
         Lazy<MetricMeter> RuntimeMetricMeter,
         IReadOnlyCollection<Type>? WorkerLevelFailureExceptionTypes,
         bool DisableEagerActivityExecution,
+        Func<PatchActivationInput, bool>? PatchActivationCallback,
         Action<string> AssertValidLocalActivity,
         VersioningBehavior? DefaultVersioningBehavior,
         WorkerDeploymentOptions? DeploymentOptions);